Floats split

+++ src/GUI/UI/UIPopup.cpp
@@ -2246,7 +2246,7 @@ void Inventory_ItemPopupAndAlchemy() {
             pAudioPlayer->playUISound(SOUND_fireBall);
             engine->_messageQueue->addMessageCurrentFrame(UIMSG_Escape, 0, 0);
 
-            Vec3i pos = pParty->pos.toInt() + Vec3i(0, 0, pParty->eyeLevel) + Vec3i::fromPolar(64, pParty->_viewYaw, pParty->_viewPitch);
+            Vec3f pos = pParty->pos + Vec3f(0, 0, pParty->eyeLevel) + Vec3f::fromPolar(64, pParty->_viewYaw, pParty->_viewPitch);
             SpriteObject::dropItemAt(SPRITE_SPELL_FIRE_FIREBALL_IMPACT, pos, 0);
             if (pParty->activeCharacter().CanAct()) {
                 pParty->activeCharacter().playReaction(SPEECH_POTION_EXPLODE);
